Include slider display and reorganize plot utilities.
