+++ mbirjax/__init__.py
@@ -1,4 +1,5 @@
-from .mbirjax import *
+from .mbir import *
 from .vcd_utils import *
 from .gpu_memory import *
+from .parallel_beam import *
